Harmonize-main/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.13.3
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
gdown>=5.1.0
gevent==24.2.1
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
huggingface-hub==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
keras==3.1.1
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv
Pygments==2.17.2
PyGithub==2.3.0
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
rsa==4.9
safetensors==0.4.2
setuptools==69.2.0
six==1.16.0
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.2
transformers==4.39.1
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.43.0
wrapt==1.16.0
zope.event==5.0
zope.interface==6.2
